+++ src/main/kotlin/com/coder/gateway/sdk/TemplateIconDownloader.kt
@@ -3,12 +3,12 @@ package com.coder.gateway.sdk
 import com.coder.gateway.icons.CoderIcons
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
-import com.intellij.ui.scale.ScaleContext
+import com.intellij.util.IconUtil
 import com.intellij.util.ImageLoader
 import com.intellij.util.ui.ImageUtil
+import org.imgscalr.Scalr
 import java.net.URL
 import javax.swing.Icon
-import javax.swing.ImageIcon
 
 @Service(Service.Level.APP)
 class TemplateIconDownloader {
@@ -25,10 +25,7 @@ class TemplateIconDownloader {
         if (url != null) {
             var img = ImageLoader.loadFromUrl(url)
             if (img != null) {
-                if (ImageUtil.getRealHeight(img) > 32 && ImageUtil.getRealWidth(img) > 32) {
-                    img = ImageUtil.resize(img, 32, ScaleContext.create())
-                }
-                return ImageIcon(img)
+                return IconUtil.toRetinaAwareIcon(Scalr.resize(ImageUtil.toBufferedImage(img), Scalr.Method.ULTRA_QUALITY, 32))
             }
         }