+++ CHANGELOG.md
@@ -4,6 +4,16 @@
 
 ## Unreleased
 
+### Added
+- ability to open a template in the Dashboard
+
+### Changed
+- renamed the plugin from `Coder Gateway` to `Gateway`
+- workspaces and agents are now resolved and displayed progressively
+
+### Fixed
+- icon rendering on macO
+
 ## 2.1.3-eap.0 - 2022-12-12
 Bug fixes and enhancements included in `2.1.3` release: